header structured, tailwind.config created, postcss installed
