+++ src/Layouts/MainLayout/MainLayout.tsx
@@ -1,10 +1,13 @@
 import { Outlet } from "react-router";
-import Navbar from "../../components/Navbar/Navbar";
+import Navbar from "../../components/Header/Navbar";
+import Header from "../../components/Header/Header";
 
 export default function MainLayout() {
   return (
     <>
+	  <Header>
       <Navbar />
+	  </Header>
       <Outlet />
     </>
   );
