+++ src/components/Header/Header.tsx
@@ -1,0 +1,11 @@
+import {  ReactNode } from "react";
+
+function Header({children}: {children: ReactNode}) {
+	return (
+		<div className="container">
+		{children}
+		</div>
+	 );
+}
+
+export default Header;
