+++ src/components/Header/SearchBar.tsx
@@ -1,0 +1,10 @@
+export const SearchBar = () => {
+	return (
+	  <input
+		type="text"
+		placeholder="Search"
+		className="input input-bordered w-24 md:w-auto input-primary"
+	  />
+	);
+  };
+
